+++ include/tags/abstract/CharacterIdTag.hpp
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <cstdint>
+
 namespace unSWF::tags
 {
     class CharacterIdTag
     {
     public:
-        virtual ~CharacterIdTag()          = default;
-        virtual int getCharacterId() const = 0;
+        virtual ~CharacterIdTag()               = default;
+        virtual uint16_t getCharacterId() const = 0;
     };
 }
